Fixed warnings as mentioned in https://dev.vatsim-germany.org/boards/22/topics/2587?r=2598#message-2598 and found by +win32-msvc*:QMAKE_CXXFLAGS_WARN_ON *= /Wall /W3 /wd4640 /wd4619 /wd4350 /wd4351

+++ src/blackgui/components/settingsfsxcomponent.cpp
@@ -69,7 +69,7 @@ namespace BlackGui
                 CLogMessage(this).warning("invalid port");
                 return;
             }
-            quint16 p = port.toUInt();
+            int p = port.toInt();
             QString msg;
             if (!CNetworkUtils::canConnect(address, p, msg))
             {
@@ -108,7 +108,7 @@ namespace BlackGui
                 CLogMessage(this).warning("invalid port");
                 return;
             }
-            quint16 p = port.toUInt();
+            int p = port.toInt();
             QString fileName = this->getIContextSimulator()->getSimulatorInfo().getSimulatorSetupValueAsString(CFsxSimulatorSetup::SetupSimConnectCfgFile);
             Q_ASSERT(!fileName.isEmpty());
             // write either local or remote file
